+++ src/config/index.js
@@ -1,0 +1,2 @@
+export { default as instance } from './axios';
+export { default as utils } from './utils';
